+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 - Allow subnormal float values in pydantic fields
+- Fix bug with inplace=False for dataframe enrichment
 
 ## [4.8.3] - 2022-08-05
 - Add repeating label to reccurence