+++ CHANGELOG.md
@@ -16,6 +16,7 @@ All notable changes to this project will be documented in this file.
 ## [4.8.2] - 2022-08-01
 - Renamed `predicted_mcc` to `mcc` for simplicity.
 - Increased mcc range to 700-9999
+- Support iterables in `add_transactions` and `add_transactions_async`
 
 ## [4.8.1] - 2022-07-22
 - Fixed a bug with deprecated `model` parameter for custom models